backend/requirements.txt
aiogram==2.11.2
aiohttp==3.7.4.post0
APScheduler==3.7.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.5
django-cors-headers==3.8.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djoser==2.1.0
environs==9.3.2
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
loguru==0.5.3
MarkupSafe==2.0.1
marshmallow==3.13.0
multidict==5.1.0
oauthlib==3.1.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.7
yarl==1.6.3
